requirements.txt
opensearch-py~=2.8.0
pydantic~=2.10.6
python-dotenv~=1.1.0